create category form. implement it to add category page. next step test its functionality
