+++ src/pages/admin/AddCategory.jsx
@@ -1,0 +1,11 @@
+import CategoryForm from "../../components/organisms/forms/CategoryForm"
+
+function AddCategory() {
+  return (
+    <>
+        <CategoryForm />
+    </>
+  )
+}
+
+export default AddCategory
