+++ src/router/router.jsx
@@ -5,6 +5,7 @@ import ProductList from "../pages/ProductList";
 import AdminProductList from "../pages/admin/AdminProductList";
 import AddProduct from "../pages/admin/AddProduct";
 import AdminCategoryList from "../pages/admin/AdminCategoryList";
+import AddCategory from "../pages/admin/AddCategory";
 
 export const router = createBrowserRouter([
     {
@@ -43,12 +44,16 @@ export const router = createBrowserRouter([
                 element: <AdminProductList />
             },               
             {
-                path: "crear",
+                path: "crear-producto",
                 element: <AddProduct />
             },
             {
                 path: "categorias",
                 element: <AdminCategoryList />
+            },
+            {
+                path: "crear-categoria",
+                element: <AddCategory />
             }
         ],
     },
